refactor(deep_causality_quantum): internal source re organization.
