+++ deep_causality_quantum/src/types/mod.rs
@@ -4,5 +4,10 @@
  */
 
 pub(crate) mod density_matrix;
+pub(crate) mod qcm;
+#[cfg(feature = "qpu")]
+pub(crate) mod qpu;
+pub(crate) mod verdict;
+pub(crate) mod qgates;
 
 pub use density_matrix::*;
