+++ deep_causality_quantum/src/types/qcm/environment.rs
