+++ deep_causality_quantum/src/types/verdict/projection.rs
@@ -18,7 +18,7 @@
 //! (see [`crate::verdict::born`]).
 
 use crate::QuantumError;
-use crate::kernels::operator_linalg::{hermiticity_defect, identity_matrix, square_dim};
+use crate::types::qgates::operator_linalg::{hermiticity_defect, identity_matrix, square_dim};
 use deep_causality_algebra::{RealField, Verdict};
 use deep_causality_num::FromPrimitive;
 use deep_causality_num_complex::Complex;
