+++ deep_causality_quantum/tests/mod.rs
@@ -3,9 +3,4 @@
  * Copyright (c) 2023 - 2026. The DeepCausality Authors and Contributors. All Rights Reserved.
  */
 mod error;
-mod kernels;
-mod qcm;
-#[cfg(feature = "qpu")]
-mod qpu;
 mod types;
-mod verdict;
